refactor(components): refactor page + form components
