+++ src/pages/admin/trilha/[id].tsx
@@ -1,11 +1,11 @@
-import { AppLayout } from 'components/AppLayout'
 import { ButtonLink } from 'components/ButtonLink'
+import { ModuleForm } from 'components/Forms/ModuleForm'
 import { Heading } from 'components/Heading'
 import { InputSearch } from 'components/Inputs/InputSearch'
-import { ModuleForm } from 'components/ModuleForm'
-import { PageFooter } from 'components/PageFooter'
-import { PageHeader } from 'components/PageHeader'
-import { SearchLoader } from 'components/SearchLoader'
+import { Loader } from 'components/Loader'
+import { Layout } from 'components/Page/Layout'
+import { PageFooter } from 'components/Page/PageFooter'
+import { PageHeader } from 'components/Page/PageHeader'
 import { Text } from 'components/Text'
 import { getCookie } from 'cookies-next'
 import { getAuthUser } from 'libs/auth/api'
@@ -58,9 +58,9 @@ export default function AdminTrailPage({
         <title>Orange Evolution | Trilha {trail?.title}</title>
       </Head>
 
-      <AppLayout>
+      <Layout>
         {isLoading || !trail ? (
-          <SearchLoader />
+          <Loader />
         ) : (
           <>
             <PageHeader
@@ -88,7 +88,7 @@ export default function AdminTrailPage({
                   />
                 </SearchWrapper>
 
-                {isSearching && <SearchLoader />}
+                {isSearching && <Loader />}
 
                 {filteredModules.length === 0 && !isSearching && (
                   <Text size="lg">Nenhum módulo encontrado 🙃</Text>
@@ -117,7 +117,7 @@ export default function AdminTrailPage({
             <PageFooter />
           </>
         )}
-      </AppLayout>
+      </Layout>
     </>
   )
 }
